Create MH Premier Template.ht5
First Version of a new HomeGauge RV MH Premier Template. It is a clean sheet re code which reduced template file size from 1672 KB to 383 KB and from 38595 lines to 10074.

It is hoped that this 87% reduction in size will make it faster, and more reliable with the HomeGauge  Companion App.